# Add ability to mix in delegate (types) to proxies