--- a/src/Castle.Core.Tests/DynamicProxy.Tests/MixinDataTestCase.cs
+++ b/src/Castle.Core.Tests/DynamicProxy.Tests/MixinDataTestCase.cs
@@ -239,5 +239,51 @@ namespace Castle.DynamicProxy.Tests
 			MixinData mixinData2 = new MixinData(new object[] { typeof(Predicate<object>) });
 			Assert.AreNotEqual(mixinData1, mixinData2);
 		}
+
+		[Test]
+		public void Ctor_succeeds_when_mixing_regular_mixin_instances_with_delegate_mixins()
+		{
+			var mixinData = new MixinData(new object[]
+			{
+				new NotADelegate(),
+				new Action(() => { }),
+			});
+		}
+
+		[Test]
+		public void Ctor_succeeds_when_mixing_regular_mixin_instances_with_delegate_type_mixins()
+		{
+			var mixinData = new MixinData(new object[]
+			{
+				new NotADelegate(),
+				typeof(Action),
+			});
+		}
+
+		[Test]
+		public void Ctor_throws_when_multiple_delegate_mixins_for_same_Invoke_signature()
+		{
+			Assert.Throws<ArgumentException>(() => new MixinData(new object[]
+			{
+				new NotADelegate(),
+				new Func<object, bool>(_ => true),
+				new Predicate<object>(_ => false),
+			}));
+		}
+
+		[Test]
+		public void Ctor_throws_when_multiple_delegate_type_mixins_for_same_Invoke_signature()
+		{
+			Assert.Throws<ArgumentException>(() => new MixinData(new object[]
+			{
+				typeof(Func<object, bool>),
+				new NotADelegate(),
+				typeof(Predicate<object>)
+			}));
+		}
+
+		public class NotADelegate : INotADelegate { }
+
+		public interface INotADelegate { }
 	}
 }