… CHANGELOG.md
… src/Castle.Core/DynamicProxy/Contributors/MixinContributor.cs
… src/Castle.Core/DynamicProxy/Generators/Emitters/ClassEmitter.cs
… src/Castle.Core/DynamicProxy/Internal/TypeUtil.cs
… src/Castle.Core/DynamicProxy/MixinData.cs
… src/Castle.Core/DynamicProxy/ProxyGenerationOptions.cs
… src/Castle.Core/DynamicProxy/ProxyUtil.cs