This is a follow-up to #403, which started out with the goal of adding a set of  `ProxyGenerator.CreateDelegateProxy` methods, and ended with us thinking about achieving delegate proxies through DynamicProxy's mixin capabilities. I am closing the earlier PR in favour of this present one.

This PR adds the following capabilities to DynamicProxy:

 * **`ProxyGenerationOptions.AddDelegateMixin(Delegate @delegate)`**  
   to add a delegate mixin. This will produce an `Invoke` method on the proxy that has the same signature as the delegate. The delegate can be `Proceed`-ed to.

 * **`ProxyGenerationOptions.AddDelegateTypeMixin(Type delegateType)`**  
   adds an `Invoke` method on the proxy that has the same signature as a specified delegate type. When those `Invoke` methods are intercepted, they cannot be proceeded from as that mixin has no target.

 * **`ProxyUtil.TryCreateDelegateToMixin(object target, Type delegateType, out Delegate @delegate)`**  
   **`ProxyUtil.TryCreateDelegateToMixin<TDelegate>(object target, out TDelegate @delegate)`**  
   are helper methods to create a delegate of a given type to a suitable `Invoke` method on a proxy.

By combining the above methods, you can now create a delegate proxy:

```csharp
var options = new ProxyGenerationOptions();
options.AddDelegateTypeMixin(typeof(Action));

var _ = generator.CreateClassProxy(typeof(object), options, ...);

ProxyUtil.TryCreateDelegateToMixin(_, out Action proxy);
                                      ^^^^^^^^^^^^^^^^
```

Because you're creating a class (or interface) proxy, you have full control over the "container" type for the generated interceptable `Invoke` method. You're also free to add custom attributes to that type, as you normally would.

I'd still like to...

* [X] ~~run some tests whether these new capabilities are actually useful in consuming libraries (Moq in my case)~~

   See [feedback for Moq](https://github.com/castleproject/Core/pull/436#issuecomment-472966476) and [NSubstitute](https://github.com/castleproject/Core/pull/436#issuecomment-473072088) below. Feedback is better than anticipated.

* [X] ~~I am also still unsure whether there may be serialization issues when using lambda delegates as mixin targets... and whether that is relevant in practice.~~

   ⚠️ If you use a lambda that captures any variables as a delegate mixin, then you won't be able to save the generated dynamic assembly to disk because the captured variables are put in a non-serializable "display class" by the compiler.

   This can be worked around by rewriting such lambdas & captures as explicit types. Beyond that, I'm not yet sure this is even fixable; if so, then we'd likely have to let `ProxyGenerationOptions` implement `ISerializable` and serialize it differently (specifically the `mixins` field).

   For now, I'd like to treat this as a known current limitation of the delegate mixin story, as it's probably not an immediate roadblock for most downstream libs.

Closes #399 (where we said that noone except DP itself should emit into its dynamic assembly, and the only known roadblock to that is delegate proxying; so if we get that, we no longer need `ModuleScope.DefineType`).

/cc @blairconrad, @thomaslevesque, @zvirja