@zvirja 2019-03-14
Hi @stakx!

The way you decided to sort it out looks promising and elegant!

I’ll definitely give it a more precise look. In the meanwhile, I want to clarify whether the following feature is supported (found the concern [in my old comment](https://github.com/castleproject/Core/pull/403#issuecomment-412677935)):

> I've just reviewed the NSubstitute lib and the usages we have and found that new API lucks a few features:
> - ...
> - ability to emit custom attributes to the generated method. We need this to quickly get the original delegate type from the methodInfo instance we internally store for the each recorded call. Basically, we use this attribute also to understand that methodInfo is a delegate proxy, rather than a method from a class proxy, so we render the client error messages more precisely.

Thanks! :)