@stakx 2019-03-14
@zvirja, nice to hear from you again! 

> ability to emit custom attributes to the generated method

I can't remember all the details from the previous issue, but basically this PR follows @jonorossi's proposal made in https://github.com/castleproject/Core/pull/403#issuecomment-422787035, where he also wrote:

> It doesn't allow you to emit custom attributes on the proxy's methods (which sounded like a nice to have, not a requirement), but it does still allow custom attributes on the proxy type [...]

That pretty much sums it up. DynamicProxy didn't have the ability to add custom attributes to single methods (AFAIK) and that hasn't changed. If that's needed, perhaps we'd need to look at it in a separate PR.