@stakx 2019-03-14
Note to self: `MixinData` equality comparison is broken when delegate type mixins are present. This will need some more work:

https://github.com/castleproject/Core/blob/a53d6b6190cde9d8b2f4e872078f769968c0ff6c/src/Castle.Core/DynamicProxy/MixinData.cs#L91-L92

https://github.com/castleproject/Core/blob/a53d6b6190cde9d8b2f4e872078f769968c0ff6c/src/Castle.Core/DynamicProxy/MixinData.cs#L112-L115

(`mixinsImpl[i]` can be null for delegate type mixins.)