@zvirja 2019-03-14
@stakx Thanks for the replies. It started to ring a bell in my head now.

Tried to give it a go, but it fails with the following error:
```
Castle.DynamicProxy.InvalidMixinConfigurationException : There is a problem with the mixins added to this ProxyGenerationOptions: The list of mixins contains at least two delegate mixins for the same delegate signature.
Parameter name: mixinInstances
  ----> System.ArgumentException : The list of mixins contains at least two delegate mixins for the same delegate signature.
Parameter name: mixinInstances
   at Castle.DynamicProxy.ProxyGenerationOptions.Initialize()
...
```

In our code I also mix in an additional object. It might be related to the finding you already described. Please ping me when it's ready, so I can give it a try 😉

From the first glance I could say that API looks perfect from the client perspective and should cover all the needs we have 👍 

One more question (warning, might be stupid): are you sure that after all the changes you applied the container generated type is still cached?