@stakx 2019-03-14
> it fails with the following error

I would need to see your code, for the moment I'll make a blind guess: Did you perhaps try to reuse the same `ProxyGenerationOptions` for different delegate proxy types? I suspect that you might end up with the same usage pattern as I did: You need at least one distinct `ProxyGenerationOptions` instance per mocked delegate type.

> are you sure that after all the changes you applied the container generated type is still cached?

I ran some quick tests this morning. Basically, caching should work as before if equality is correctly implemented on both `ProxyGenerationOptions` (and `MixinData`, which is used by the former), I've added some additional unit tests for that. You can test for yourself by attaching e.g. a `ConsoleLogger` to your `ProxyGenerator`. When you ask for a proxy, you should see a log message stating whether a cached type was found (and used) or not.