@stakx 2019-03-14
@zvirja, there was a glaringly obvious bug that I didn't notice anymore at 2am last night. 😁 Fixing it also fixed all NSubstitute unit tests except one (`When_checking_across_multiple_subs_including_delegates`, which appears to fail because of differently named types).