@stakx 2019-03-14
Speaking for Moq, I am very satisfied with how well this change would work out: https://github.com/stakx/moq4/commit/50fd9fa53e4b3af2e431e7fdbddce8f57b034c04 (14 insertions, 122 deletions). Almost all of the delegate special-casing code is gone, and gets replaced by the very short `ProxyUtil.TryCreateDelegateToMixin` bit.