@zvirja 2019-03-14
@stakx After working a bit more on our code I managed to finally get it work. Because `MethodInfo` passed to the invocation is actually taken from the delegate type now, I don't even need to emit custom attributes  - simply use `methodInfo.DeclaringType` to get original delegate type 😊 I really enjoy the way it goes! Thanks a lot! 🤝 

The only issue I discovered is that delegate proxy activation degraded in terms of performance:

```
BenchmarkDotNet=v0.10.14, OS=Windows 10.0.17134
Intel Core i7-8850H CPU 2.60GHz (Coffee Lake), 1 CPU, 12 logical and 6 physical cores
Frequency=2531252 Hz, Resolution=395.0614 ns, Timer=TSC
.NET Core SDK=2.1.504
  [Host] : .NET Core 2.1.8 (CoreCLR 4.6.27317.03, CoreFX 4.6.27317.03), 64bit RyuJIT
  Clr    : .NET Framework 4.7.1 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.3324.0
  Core   : .NET Core 2.1.8 (CoreCLR 4.6.27317.03, CoreFX 4.6.27317.03), 64bit RyuJIT

BEFORE

                           Method |  Job | Runtime |      Mean |     Error |    StdDev |
--------------------------------- |----- |-------- |----------:|----------:|----------:|
         CreateDelegateSubstitute |  Clr |     Clr | 10.837 us | 0.1587 us | 0.1407 us |
         CreateDelegateSubstitute | Core |    Core |  7.557 us | 0.0533 us | 0.0352 us |
		 
AFTER

                           Method |  Job | Runtime |      Mean |     Error |    StdDev |
--------------------------------- |----- |-------- |----------:|----------:|----------:|
         CreateDelegateSubstitute |  Clr |     Clr | 13.452 us | 0.1607 us | 0.1424 us |
         CreateDelegateSubstitute | Core |    Core |  9.630 us | 0.1893 us | 0.2714 us |
```

Obviously, I don't want you to profile our code 😅 But if you could benchmark Moq instead, it would be interesting to see whether you notice the same degradation. So it might be easier for you to check whether degradation resides somewhere in the code you are adding (as previously we also used `Castle.Proxy` to activate proxies and `Delegate.CreateDelegate()` to make delegates)...🤔